Safe engine.execute for zipcode lookup
This was hard-coded, which is a lot better than an injection site, but I
want my zip code to show up :)

I'm using sqlalchemy.sql.text() to give us parameterized and
type-validated input even though we're not using a model here.

Returning None for invalid data bubbles up to a 400 error which seems
appropriate to me.

Signed-off-by: Josh Farwell <josh.farwell@gmail.com>